Make rain rate variable and remove randomness/dithering. GUI still needed
With the higher bit depth the rain dithering is not necessary anymore.
It did look cool, but for the cool looks it would be better to add an effect
in visualisation.

to set rain rate from console:
>> raining=x
with x = rain rate in m per simulation step. Default: 0.001, 1mm per simulation step